completed User Authenticated requests (get/create/update/delete contact)

+++ controllers/contactController.js
@@ -2,15 +2,15 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req,res) =>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc Create new contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req,res) =>{
     console.log("The request body is:", req.body);
     const {name,email,phone} = req.body;
@@ -19,31 +19,39 @@ const createContact = asyncHandler(async (req,res) =>{
         throw new Error("All fields are mandatory!");
     }
     const contact = await Contact.create({
-        name, email, phone,
+        name, email, phone, user_id:req.user.id
     });
     res.status(201).json(contact);
 });
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
+    if(contact.user_id.toString() != req.user.id){
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user's contacts");
+    }
     res.status(200).json(contact);
 });
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user's contacts");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -55,12 +63,16 @@ const updateContact = asyncHandler(async (req,res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user's contacts");
     }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
